+++ lact-schema/src/tests.rs
@@ -1,4 +1,4 @@
-use crate::{FanControlMode, FanOptions, PmfwOptions, Pong, Request, Response};
+use crate::{FanControlMode, FanOptions, PmfwOptions, Pong, Request, Response, clean_gpu_name};
 use anyhow::anyhow;
 use serde_json::json;
 use std::collections::BTreeMap;
@@ -85,4 +85,49 @@ fn set_fan_clocks() {
         change_threshold: None,
     });
     assert_eq!(expected_request, request);
+}
+
+#[test]
+fn clean_gpu_name_removes_vendor_prefixes() {
+    assert_eq!(clean_gpu_name("AMD Radeon RX 9070 XT"), "RX 9070 XT");
+    assert_eq!(clean_gpu_name("NVIDIA GeForce RTX 5090"), "RTX 5090");
+    assert_eq!(clean_gpu_name("NVIDIA GeForce MX450"), "MX450");
+    assert_eq!(
+        clean_gpu_name("NVIDIA GeForce RTX 5090 [Founders Edition]"),
+        "RTX 5090"
+    );
+    assert_eq!(
+        clean_gpu_name("NVIDIA GeForce RTX 4070 Super"),
+        "RTX 4070 Super"
+    );
+    assert_eq!(clean_gpu_name("Intel Arc A380"), "Arc A380");
+}
+
+#[test]
+fn clean_gpu_name_unwraps_pci_names() {
+    assert_eq!(
+        clean_gpu_name("Pitcairn XT [Radeon HD 7870 GHz Edition]"),
+        "HD 7870 GHz Edition"
+    );
+    assert_eq!(clean_gpu_name("DG2 [Arc A380]"), "Arc A380");
+    assert_eq!(clean_gpu_name("GK107M [GeForce 710A]"), "710A");
+    assert_eq!(clean_gpu_name("GK107M [GeForce 820M]"), "820M");
+    assert_eq!(clean_gpu_name("TU117M [GeForce MX450]"), "MX450");
+    assert_eq!(
+        clean_gpu_name("TigerLake-LP GT2 [Iris Xe Graphics]"),
+        "Iris Xe Graphics"
+    );
+}
+
+#[test]
+fn clean_gpu_name_strips_consumer_brands_and_keeps_unrecognized_names() {
+    assert_eq!(clean_gpu_name("AMD Radeon 780M Graphics"), "780M Graphics");
+    assert_eq!(clean_gpu_name("Phoenix1"), "Phoenix1");
+}
+
+#[test]
+fn clean_gpu_name_keeps_professional_product_brands() {
+    assert_eq!(clean_gpu_name("GK107GL [Quadro K600]"), "Quadro K600");
+    assert_eq!(clean_gpu_name("GK110GL [Tesla K20]"), "Tesla K20");
+    assert_eq!(clean_gpu_name("NVIDIA Quadro RTX 6000"), "Quadro RTX 6000");
 }
